backend/requirements.txt
asgiref==3.11.1
charset-normalizer==3.4.4
Django==5.1
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.1.1
PyJWT==2.11.0
reportlab==4.4.10
sqlparse==0.5.5
